Folder Structure improved
