+++ cf/java/_2024_12_09/foxSnake.java
@@ -1,5 +1,6 @@
-package decPast;
-import java.util.*;
+package _2024_12_09;
+
+import java.util.Scanner;
 
 public class foxSnake {
     public static void main(String[] args)
