+++ cf/java/_2024_12_09/generalSquad.java
@@ -1,4 +1,4 @@
-package decPast;
+package _2024_12_09;
 import java.util.*;
 
 public class generalSquad {
